Bug 1682764 - Reorganize cycling components
